Set a color for "ghost_element.disabled"
I haven't been able to test this color, so I'm not really sure how it
will look.  Currently, it only seems to be used for a few Extensions
view buttons, mostly during transient states that are hard to simulate
(e.g. "Uninstall" and "Upgrade" buttons are disabled while an update is
being applied).